+++ nodes/manageErrors.js
@@ -1,6 +1,7 @@
 
 function VerifyErrorSL (node, msg, response, consentEmpty=false) {
     if (!response.hasOwnProperty("success") && !response.hasOwnProperty("error") && !Object.keys(response).length && !consentEmpty) {  // Error Generic
+        
         msg.payload = response;
         node.error('Not Valid Requests', msg)
         node.status({ fill: 'red', shape: 'dot', text: 'Not Valid  Requests' });
@@ -25,8 +26,9 @@ function VerifyErrorSL (node, msg, response, consentEmpty=false) {
         }
 
     }
-    else { // OK Response 
+    else { // OK Response
         return response;
+        
     }