+++ nodes/manipulateEntitySap.js
@@ -39,7 +39,7 @@ module.exports = function (RED) {
   }
 
   RED.httpAdmin.get('/entities', RED.auth.needsPermission('manipulateEntitySap.read'), (req, res) => {
-    console.log('entities');
+    //console.log('entities');
     res.json(entities);
   });
   RED.nodes.registerType('manipulateEntitySap', ManipulateEntitySap, {});